requirements.txt
Flask
Flask-Mail
python-dotenv
psycopg2-binary
Werkzeug
click
